removed unused duplicate ParameterPanel.tsx component and move LandingPage to pages
